+++ public_html/app/Services/Organization/OrganizationalContextException.php
@@ -1,0 +1,3 @@
+<?php
+namespace App\Services\Organization;
+class OrganizationalContextException extends \RuntimeException {}